requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.8.4
asyncio>=3.4.3
requests>=2.31.0
typing-extensions>=4.7.0
chainlit>=2.6.2
